tests: added base class for setup/teardown database before/after each method call
